requirements.txt
Flask==0.10.1
Flask-Migrate==1.2.0
Flask-SQLAlchemy==1.0
Flask-Script==2.0.3
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.21
SQLAlchemy==0.9.4
Werkzeug==0.9.4
alembic==0.6.5
argparse==1.2.1
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
gunicorn==18.0
itsdangerous==0.24
psycopg2==2.5.2
wsgiref==0.1.2
